# src/app/schemas/common.py
"""
Esquemas comunes. Origina en: contrato-canales-fuentes-rss.openapi.yaml,
componentes `Error`.
"""
from typing import List, Optional

from pydantic import BaseModel


class Error(BaseModel):
    codigo: str
    mensaje: str
    detalles: Optional[List[str]] = None